requirements.txt
python-telegram-bot==21.4
supabase-py==2.3.5
requests==2.31.0
python-dotenv==1.0.1
